fix(error)!: Consolidate Parser combinators
