+++ src/combinator/mod.rs
@@ -152,7 +152,7 @@
 //! - [`space0`][crate::character::space0]: Recognizes zero or more spaces and tabs. [`space1`][crate::character::space1] does the same but returns at least one character
 //! - [`tab`][crate::character::tab]: Matches a tab character `\t`
 
-use crate::error::{ErrMode, ErrorKind, FromExternalError, Needed, ParseError};
+use crate::error::{ContextError, ErrMode, ErrorKind, FromExternalError, Needed, ParseError};
 use crate::input::Offset;
 use crate::input::{Input, Location};
 use crate::lib::std::borrow::Borrow;
@@ -1592,4 +1592,79 @@ pub fn success<I, O: Clone, E: ParseError<I>>(val: O) -> impl Fn(I) -> IResult<I
 /// ```
 pub fn fail<I, O, E: ParseError<I>>(i: I) -> IResult<I, O, E> {
   Err(ErrMode::Backtrack(E::from_error_kind(i, ErrorKind::Fail)))
+}
+
+/// Implementation of [`Parser::context`]
+#[cfg_attr(nightly, warn(rustdoc::missing_doc_code_examples))]
+pub struct Context<F, O, C: Clone> {
+  f: F,
+  context: C,
+  phantom: core::marker::PhantomData<O>,
+}
+
+impl<F, O, C: Clone> Context<F, O, C> {
+  pub(crate) fn new(f: F, context: C) -> Self {
+    Self {
+      f,
+      context,
+      phantom: Default::default(),
+    }
+  }
+}
+
+impl<I, O, E, F, C> Parser<I, O, E> for Context<F, O, C>
+where
+  I: Clone,
+  C: Clone,
+  E: ContextError<I, C>,
+  F: Parser<I, O, E>,
+{
+  fn parse_next(&mut self, i: I) -> IResult<I, O, E> {
+    (self.f)
+      .parse_next(i.clone())
+      .map_err(|err| err.map(|err| err.add_context(i, self.context.clone())))
+  }
+}
+
+/// Implementation of [`Parser::dbg_err`]
+#[cfg_attr(nightly, warn(rustdoc::missing_doc_code_examples))]
+#[cfg(feature = "std")]
+pub struct DbgErr<F, O, C> {
+  f: F,
+  context: C,
+  phantom: core::marker::PhantomData<O>,
+}
+
+#[cfg(feature = "std")]
+impl<F, O, C> DbgErr<F, O, C> {
+  pub(crate) fn new(f: F, context: C) -> Self {
+    Self {
+      f,
+      context,
+      phantom: Default::default(),
+    }
+  }
+}
+
+#[cfg(feature = "std")]
+impl<I, O, E, F, C> Parser<I, O, E> for DbgErr<F, O, C>
+where
+  I: crate::input::AsBytes,
+  I: Clone,
+  E: std::fmt::Debug,
+  F: Parser<I, O, E>,
+  C: std::fmt::Display,
+{
+  fn parse_next(&mut self, input: I) -> IResult<I, O, E> {
+    use crate::input::HexDisplay;
+    let i = input.clone();
+    match self.f.parse_next(i) {
+      Err(e) => {
+        let input = input.as_bytes();
+        eprintln!("{}: Error({:?}) at:\n{}", self.context, e, input.to_hex(8));
+        Err(e)
+      }
+      a => a,
+    }
+  }
 }
