+++ src/error.rs
@@ -975,41 +975,6 @@ where
   }
 }
 
-/// Implementation of [`Parser::context`]
-#[cfg_attr(nightly, warn(rustdoc::missing_doc_code_examples))]
-pub struct Context<F, O, C: Clone> {
-  f: F,
-  context: C,
-  phantom: core::marker::PhantomData<O>,
-}
-
-impl<F, O, C: Clone> Context<F, O, C> {
-  pub(crate) fn new(f: F, context: C) -> Self {
-    Self {
-      f,
-      context,
-      phantom: Default::default(),
-    }
-  }
-}
-
-impl<I, O, E, F, C> Parser<I, O, E> for Context<F, O, C>
-where
-  I: Clone,
-  C: Clone,
-  E: ContextError<I, C>,
-  F: Parser<I, O, E>,
-{
-  fn parse_next(&mut self, i: I) -> IResult<I, O, E> {
-    match (self.f).parse_next(i.clone()) {
-      Ok(o) => Ok(o),
-      Err(ErrMode::Incomplete(i)) => Err(ErrMode::Incomplete(i)),
-      Err(ErrMode::Backtrack(e)) => Err(ErrMode::Backtrack(e.add_context(i, self.context.clone()))),
-      Err(ErrMode::Cut(e)) => Err(ErrMode::Cut(e.add_context(i, self.context.clone()))),
-    }
-  }
-}
-
 /// Transforms a `VerboseError` into a trace with input position information
 #[cfg(feature = "alloc")]
 pub fn convert_error<I: core::ops::Deref<Target = str>>(
@@ -1279,49 +1244,6 @@ where
       Err(e)
     }
     a => a,
-  }
-}
-
-/// Implementation of [`Parser::dbg_err`]
-#[cfg_attr(nightly, warn(rustdoc::missing_doc_code_examples))]
-#[cfg(feature = "std")]
-pub struct DbgErr<F, O, C> {
-  f: F,
-  context: C,
-  phantom: core::marker::PhantomData<O>,
-}
-
-#[cfg(feature = "std")]
-impl<F, O, C> DbgErr<F, O, C> {
-  pub(crate) fn new(f: F, context: C) -> Self {
-    Self {
-      f,
-      context,
-      phantom: Default::default(),
-    }
-  }
-}
-
-#[cfg(feature = "std")]
-impl<I, O, E, F, C> Parser<I, O, E> for DbgErr<F, O, C>
-where
-  I: crate::input::AsBytes,
-  I: Clone,
-  E: std::fmt::Debug,
-  F: Parser<I, O, E>,
-  C: std::fmt::Display,
-{
-  fn parse_next(&mut self, input: I) -> IResult<I, O, E> {
-    use crate::input::HexDisplay;
-    let i = input.clone();
-    match self.f.parse_next(i) {
-      Err(e) => {
-        let input = input.as_bytes();
-        eprintln!("{}: Error({:?}) at:\n{}", self.context, e, input.to_hex(8));
-        Err(e)
-      }
-      a => a,
-    }
   }
 }
